chora/requirements.txt
2captcha-python==1.2.5
absl-py==2.1.0
aenum==2.2.6
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==19.1.0
Automat==22.10.0
autoscraper==1.1.14
beautifulsoup4==4.12.3
black==24.4.2
boto3==1.34.116
botocore==1.34.116
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
collector==0.4.6
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dask==2024.5.1
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distributed==2024.5.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fbchat==1.9.7
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.5.4
gevent==24.2.1
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
googlesearch-python==1.2.3
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
heyvi==0.2.28
httmock==1.4.0
http-request-randomizer==1.3.2
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
itemadapter==0.8.0
itemloaders==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lightning-utilities==0.11.2
locket==1.0.0
lxml==5.2.1
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
maxminddb==2.6.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauth2==1.9.0.post1
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.2
parse==1.20.1
parsel==1.9.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pefile==2023.2.7
pgesmd-self-access==0.0.6
pillow==10.3.0
pip-system-certs==4.0
platformdirs==4.2.2
playwright==1.44.0
playwright-stealth==1.0.6
prompt-toolkit==3.0.43
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
proxy-pool==0.0.3
proxybroker==0.3.2
proxyscrape==0.3.0
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycares==4.4.0
pycollector==0.4.6
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
pydub==0.25.1
pyee==11.1.0
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
pyppeteer==2.0.0
pyppeteer-stealth==2.7.4
pyquery==2.0.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.5
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
requests-html==0.10.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
Scrapy==2.11.1
selenium==4.19.0
service-identity==24.1.0
setuptools-scm==8.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.3
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tldextract==5.1.2
toolz==0.12.1
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
vipy==1.14.4
w3lib==2.1.2
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==10.4
Werkzeug==3.0.3
whisper==1.1.10
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.0
zope.event==5.0
zope.interface==6.3